+++ components/Favourites/Favourites.test.js
@@ -1,20 +1,21 @@
 import Favourites from ".";
 import { render, screen } from "@testing-library/react";
-import useStore from "@/Stores/useStore";
-import useLocalStore from "@/Stores/useLocalStore";
+import useSetUpMock from "@/hooks/useSetUpMock";
 
 jest.mock("../../Stores/useStore");
 jest.mock("../../Stores/useLocalStore");
 
 test("renders all art favourite pieces as a list", () => {
-  renderFavourites();
+  useSetUpMock();
+  render(<Favourites />);
 
   expect(screen.getAllByRole("listitem")).toHaveLength(1);
   expect(screen.getByText(/The Creation of Adam/)).toBeInTheDocument();
 });
 
 test("Each art piece's image, artist and name is displayed", () => {
-  renderFavourites();
+  useSetUpMock();
+  render(<Favourites />);
 
   const image = screen.getByRole("img");
   expect(image).toBeInTheDocument();
@@ -27,42 +28,8 @@ test("Each art piece's image, artist and name is displayed", () => {
 });
 
 test("a FavoriteButton component is rendered", () => {
-  renderFavourites();
+  useSetUpMock();
+  render(<Favourites />);
 
   expect(screen.getByRole("button")).toBeInTheDocument();
 });
-
-function renderFavourites() {
-  useStore.mockReturnValueOnce([
-    {
-      slug: "the-creation-of-adam",
-      name: "The Creation of Adam",
-      artist: "Michelangelo",
-      imageSource: "/images/the-creation-of-adam.jpg",
-    },
-    {
-      slug: "the-scream",
-      name: "The Scream",
-      artist: "Edvard Munch",
-      imageSource: "/images/the-scream.jpg",
-    },
-  ]);
-
-  useLocalStore.mockImplementation((selector) => {
-    const state = {
-      artPiecesInfo: [
-        {
-          slug: "the-creation-of-adam",
-          isFavourite: true,
-        },
-        {
-          slug: "the-scream",
-          isFavourite: false,
-        },
-      ],
-    };
-    return selector(state); // Return the selected part of the state
-  });
-
-  render(<Favourites />);
-}
